test(file): replace globals in file tests
