+++ test/os/file/mime/text.test.js
@@ -1,10 +1,16 @@
 goog.require('os.file.File');
+goog.require('os.file.mime');
 goog.require('os.file.mime.mock');
 goog.require('os.file.mime.text');
 
 describe('os.file.mime.text', function() {
+  const mime = goog.module.get('os.file.mime');
+  const text = goog.module.get('os.file.mime.text');
+
+  const mockMime = goog.module.get('os.file.mime.mock');
+
   it('should detect files that are text files', function() {
-    os.file.mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/plugin/file/geojson/10k.json',
       '/base/test/plugin/file/kml/kml_test.xml',
       '/base/test/resources/text/big5.bin',
@@ -25,14 +31,14 @@ describe('os.file.mime.text', function() {
       '/base/test/resources/text/utf8.bin',
       '/base/test/resources/text/utf8.truncated-multibyte.bin',
       '/base/test/resources/text/windows-1252.french.bin',
-      '/base/test/resources/text/windows-1252.german.bin'],
-        function(buffer) {
-          expect(os.file.mime.text.getText(buffer)).toBeTruthy();
-        });
+      '/base/test/resources/text/windows-1252.german.bin'
+    ], function(buffer) {
+      expect(text.getText(buffer)).toBeTruthy();
+    });
   });
 
   it('should register itself with mime detection', function() {
-    var chain = os.file.mime.getTypeChain('text/plain').join(', ');
+    var chain = mime.getTypeChain('text/plain').join(', ');
     expect(chain).toBe('application/octet-stream, text/plain');
   });
 });
