+++ test/os/file/mime/zip.test.js
@@ -1,29 +1,35 @@
 goog.require('os.file.File');
+goog.require('os.file.mime');
 goog.require('os.file.mime.mock');
 goog.require('os.file.mime.zip');
 
 describe('os.file.mime.zip', function() {
+  const mime = goog.module.get('os.file.mime');
+  const mimeZip = goog.module.get('os.file.mime.zip');
+
+  const mockMime = goog.module.get('os.file.mime.mock');
+
   it('should not detect files that are not zip files', function() {
-    os.file.mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/resources/xml/namespaced-root-partial.xml',
       '/base/test/resources/xml/comment-with-embedded-xml.xml',
       '/base/test/plugin/file/kml/kml_test.xml',
       '/base/test/plugin/file/geojson/10k.json',
       '/base/test/resources/bin/rand.bin'],
     function(buffer) {
-      expect(os.file.mime.zip.isZip(buffer)).toBe(false);
+      expect(mimeZip.isZip(buffer)).toBe(false);
     });
   });
 
   it('should detect files that are zip files even if it only has a partial', function() {
-    os.file.mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/resources/zip/test.zip',
       '/base/test/resources/zip/test.docx',
       '/base/test/resources/zip/test.kmz'],
     function(buffer, filename) {
       var result = Number.POSITIVE_INFINITY;
       runs(function() {
-        os.file.mime.zip.detectZip(buffer).then(function(val) {
+        mimeZip.detectZip(buffer).then(function(val) {
           result = val;
         });
       });
@@ -39,13 +45,13 @@ describe('os.file.mime.zip', function() {
   });
 
   it('should detect files that are zip files and return the entries if given a complete file', function() {
-    os.file.mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/resources/zip/test.zip',
       '/base/test/resources/zip/test.kmz'],
     function(buffer, filename) {
       var result = Number.POSITIVE_INFINITY;
       runs(function() {
-        os.file.mime.zip.detectZip(buffer).then(function(val) {
+        mimeZip.detectZip(buffer).then(function(val) {
           result = val;
         });
       });
@@ -62,7 +68,7 @@ describe('os.file.mime.zip', function() {
   });
 
   it('should register itself with mime detection', function() {
-    var chain = os.file.mime.getTypeChain(os.file.mime.zip.TYPE).join(', ');
+    var chain = mime.getTypeChain(mimeZip.TYPE).join(', ');
     expect(chain).toBe('application/octet-stream, application/zip');
   });
 });
