+++ test/plugin/ogc/mime.test.js
@@ -6,23 +6,26 @@ goog.require('plugin.ogc.mime');
 describe('plugin.ogc.mime', function() {
   const mime = goog.module.get('os.file.mime');
   const ogc = goog.module.get('os.ogc');
+
+  const mockMime = goog.module.get('os.file.mime.mock');
+
   it('should not detect files that are not OGC Server files', function() {
-    mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/plugin/file/kml/kml_test.xml',
       '/base/test/resources/bin/rand.bin',
       '/base/test/resources/json/partial_object.json'
-    ], mime.mock.testNo(ogc.ID));
+    ], mockMime.testNo(ogc.ID));
   });
 
   it('should detect files that are OGC Server files', function() {
-    mime.mock.testFiles([
+    mockMime.testFiles([
       '/base/test/resources/ogc/wms-130.xml',
       '/base/test/resources/ogc/wms-111.xml',
       '/base/test/resources/ogc/wmts-100.xml',
       '/base/test/resources/ogc/wfs-200.xml', // we don't support WFS 2.0.0 but we do detect it as an OGC server
       '/base/test/resources/ogc/wfs-110.xml',
       '/base/test/resources/ogc/exception-report.xml'
-    ], mime.mock.testYes(ogc.ID));
+    ], mockMime.testYes(ogc.ID));
   });
 
   it('should register itself with mime detection', function() {
